--- composite/timeline/src/test/java/br/com/orcinus/orca/composite/timeline/figure/gallery/GalleryTests.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright © 2023-2024 Orcinus
- *
- * This program is free software: you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
- * the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program. If
- * not, see https://www.gnu.org/licenses.
- */
-
-package br.com.orcinus.orca.composite.timeline.figure.gallery
-
-import br.com.orcinus.orca.composite.timeline.post.figure.Figure
-import br.com.orcinus.orca.composite.timeline.post.figure.gallery.GalleryPreview
-import kotlin.test.Test
-
-internal class GalleryTests {
-  @Test(expected = IllegalArgumentException::class)
-  fun throwsWhenNoAttachmentsAreProvided() {
-    Figure.Gallery(GalleryPreview.sample.copy(attachments = emptyList()))
-  }
-}
