+++ composite/timeline/src/test/java/br/com/orcinus/orca/composite/timeline/post/figure/gallery/thumbnail/test/SemanticsNodeInteraction.extensions.kt
